requirements.txt
fastapi
uvicorn
pydantic-settings
openai
langchain
python-dotenv
ruff
numpy
Pillow
opencv-python
mediapipe
ultralytics
python-multipart